[refactor] auth 모듈 분리 및 전반적인 패키지 리팩토링
