+++ core/authentication/src/main/kotlin/com/ssavice/authentication/di/AuthenticationModule.kt
@@ -1,7 +1,7 @@
-package com.ssavice.network.di
+package com.ssavice.authentication.di
 
+import com.ssavice.authentication.repositoryimpl.demo.DemoAuthenticationRepository
 import com.ssavice.network.authentication.AuthenticationRepository
-import com.ssavice.network.authentication.demo.DemoAuthenticationRepository
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -9,7 +9,7 @@ import dagger.hilt.components.SingletonComponent
 
 @Module
 @InstallIn(SingletonComponent::class)
-internal interface DemoNetworkModule {
+internal interface AuthenticationModule {
     @Binds
     fun bindAuthenticationRepository(impl: DemoAuthenticationRepository): AuthenticationRepository
 }
