+++ core/authentication/src/main/kotlin/com/ssavice/authentication/repositoryimpl/demo/DemoAuthenticationRepository.kt
@@ -1,9 +1,8 @@
-package com.ssavice.network.authentication.demo
+package com.ssavice.authentication.repositoryimpl.demo
 
 import com.ssavice.model.TimeStamp
 import com.ssavice.model.auth.Jwt
 import com.ssavice.network.authentication.AuthenticationRepository
-import retrofit2.Response
 import javax.inject.Inject
 
 class DemoAuthenticationRepository
@@ -11,8 +10,8 @@ class DemoAuthenticationRepository
 
     override suspend fun refreshToken(jwt: Jwt): Jwt? = null
 
-    override suspend fun userLoginWithAccessToken(accessToken: String): Response<Jwt> {
-        return Response.success(
+    override suspend fun userLoginWithAccessToken(accessToken: String): Result<Jwt> {
+        return Result.success(
             Jwt(
                 accessToken = "token",
                 refreshToken = "refreshToken",
@@ -21,8 +20,8 @@ class DemoAuthenticationRepository
         )
     }
 
-    override suspend fun companyLoginWithAccessToken(accessToken: String): Response<Jwt> {
-        return Response.success(
+    override suspend fun companyLoginWithAccessToken(accessToken: String): Result<Jwt> {
+        return Result.success(
             Jwt(
                 accessToken = "token",
                 refreshToken = "refreshToken",
