+++ core/data/src/main/java/com/ssavice/data/di/CompanyRetrofitModule.kt
@@ -4,7 +4,7 @@ import com.ssavice.data.repository.SellerInfoRepository
 import com.ssavice.data.repositoryimpl.DemoSellerInfoRepository
 import com.ssavice.data.repositoryimpl.RemoteSellerInfoRepository
 import com.ssavice.data.service.CompanyRetrofitService
-import com.ssavice.network.retrofit.RetrofitModule
+import com.ssavice.network.di.RetrofitModule
 import dagger.Binds
 import dagger.Module
 import dagger.Provides
