+++ core/data/src/main/java/com/ssavice/data/di/ServiceRetrofitModule.kt
@@ -4,7 +4,7 @@ import com.ssavice.data.repository.ServiceRepository
 import com.ssavice.data.repositoryimpl.DemoServiceRepository
 import com.ssavice.data.repositoryimpl.RemoteServiceRepository
 import com.ssavice.data.service.ServiceRetrofitService
-import com.ssavice.network.retrofit.RetrofitModule
+import com.ssavice.network.di.RetrofitModule
 import dagger.Binds
 import dagger.Module
 import dagger.Provides
