+++ settings.gradle.kts
@@ -44,3 +44,4 @@ include(":feature:edit-profile")
 include(":feature:user-my-service")
 include(":core:datastore")
 include(":core:addresspicker")
+include(":core:authentication")
